The code is divided into translation units
