+++ mihalchenko.vladislav/T2/scopeGuard.hpp
@@ -1,0 +1,22 @@
+#ifndef SCOPEGUARD_HPP
+#define SCOPEGUARD_HPP
+
+#include <iostream>
+
+namespace mihalchenko
+{
+  class iofmtguard
+  {
+  public:
+    iofmtguard(std::basic_ios<char> &s);
+    ~iofmtguard();
+
+  private:
+    std::basic_ios<char> &s_;
+    char fill_;
+    std::streamsize precision_;
+    std::basic_ios<char>::fmtflags fmt_;
+  };
+}
+
+#endif
